Rename unified-logs-query to unified-logs-infinite-query
